% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRN3rmRedun_sameWithMetdna.R
\name{assignConfidenceInMrn3}
\alias{assignConfidenceInMrn3}
\title{assignConfidenceInMrn3}
\usage{
assignConfidenceInMrn3(result_df)
}
\arguments{
\item{result_df}{annotation result data frame with column: "tag" & "adduct" ! (same peak group!)}
}
\value{
"confidence" of this peak group (a value vector)
}
\description{
assignConfidenceInMrn3
}
\examples{
assignConfidenceInMrn3(result_df)

}
