% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRN3rmRedun_sameWithMetdna.R
\name{calculateRedundancyInMrn3}
\alias{calculateRedundancyInMrn3}
\title{calculateRedundancyInMrn3}
\usage{
calculateRedundancyInMrn3(result_df)
}
\arguments{
\item{result_df}{annotation result data frame (all annotation results!)}
}
\value{
"Peak Redundancy" & "Metabolite Redundancy" (two value vector)
}
\description{
calculateRedundancyInMrn3
}
\examples{
calculateRedundancyInMrn3(result_df)

}
