% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRN3data.R
\name{convertSpectraData}
\alias{convertSpectraData}
\title{convertSpectraData}
\usage{
convertSpectraData(ms2_data)
}
\arguments{
\item{ms2_data}{ms2_data}
}
\value{
SpectraData
}
\description{
convertSpectraData
}
\examples{
convertSpectraData(ms2_data)
}
