% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTags2ForMetDNA.R
\name{createTags2ForMetDNA}
\alias{createTags2ForMetDNA}
\title{createTags2ForMetDNA}
\usage{
createTags2ForMetDNA(ms1_data, result_df)
}
\arguments{
\item{ms1_data}{A tibble containing the MS1 data in MRN3 format with columns 'name', 'mz', 'rt', and 'ccs'.}

\item{result_df}{A table of MRN3 annotation result ("result_final.csv")}
}
\value{
Tags2 of MetDNA
}
\description{
createTags2ForMetDNA
}
\examples{
createTags2ForMetDNA(ms1_data = ms1_data, result_df = result_df)

}
