% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRN3annotation.R
\name{getAdductAnno}
\alias{getAdductAnno}
\title{getAdductAnno}
\usage{
getAdductAnno(table_anno, idx_seeds, round, rt_tol_adductAnno = 3)
}
\arguments{
\item{table_anno}{table_mrn_anno (total)}

\item{idx_seeds}{idx_seeds}

\item{round}{round}

\item{rt_tol_adductAnno}{rt_tol_adductAnno = 3 s}
}
\value{
table_mrn_anno (after adduct annotation)
}
\description{
getAdductAnno
}
\examples{
table_mrn_anno <- getAdductAnno(table_anno = table_mrn_anno, idx_seeds = c(1, 2), round = 1, rt_tol_adductAnno = 3)

}
