% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRN3data.R
\name{getErrorAndScore}
\alias{getErrorAndScore}
\title{getErrorAndScore}
\usage{
getErrorAndScore(
  table_anno,
  is_rt_valid = TRUE,
  is_ccs_valid = FALSE,
  mz_ppm_thr = 0,
  mz_tol = 15,
  rt_tol_rel = 30,
  ccs_tol_rel = 4
)
}
\arguments{
\item{table_anno}{table_anno}

\item{is_rt_valid}{is_rt_valid (TRUE)}

\item{is_ccs_valid}{is_ccs_valid (FALSE)}

\item{mz_ppm_thr}{mz_ppm_thr (0 Da for Orbitrap; 400 Da for TOF)}

\item{mz_tol}{mz_tol (MS1, 15 ppm)}

\item{rt_tol_rel}{rt_tol_rel (30 pctg.)}

\item{ccs_tol_rel}{ccs_tol_rel (4 pctg.)}
}
\value{
table_anno
}
\description{
getErrorAndScore
}
\examples{
getErrorAndScore(table_anno)
}
