% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRN3data.R
\name{getMatchParam}
\alias{getMatchParam}
\title{getMatchParam}
\usage{
getMatchParam(
  mz_tol_ms2 = 25,
  mz_ppm_thr = 400,
  includePrecursor = FALSE,
  ppmPrecursorFilter = 30,
  thrIntensityAbs = 50,
  thrIntensityRel = 0.01,
  methodScore = c("dp", "hybrid")
)
}
\arguments{
\item{mz_tol_ms2}{mz_tol_ms2 25 ppm}

\item{mz_ppm_thr}{mz_ppm_thr: 400 Da in TOF, 0 in Orbitrap}

\item{includePrecursor}{FALSE in MRN annotation}

\item{ppmPrecursorFilter}{ppmPrecursorFilter 30 ppm}

\item{thrIntensityAbs}{thrIntensityAbs 50}

\item{thrIntensityRel}{thrIntensityRel 0.01 = 1 pctg.}

\item{methodScore}{methodScore: 'dp' or 'hybrid'}
}
\value{
SpectraTools::MatchParam
}
\description{
getMatchParam
}
\examples{
getMatchParam(mz_tol_ms2 = 25, mz_ppm_thr = 0)
}
