% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRN3annotation.R
\name{getMetAnno}
\alias{getMetAnno}
\title{getMetAnno}
\usage{
getMetAnno(
  table_anno,
  idx_seeds,
  round,
  obj_mrn,
  table_ms2_network_filter,
  step_mrn = 1,
  topN_neighbor = Inf,
  lib_type_mrn = "MRN3",
  thread = 4
)
}
\arguments{
\item{table_anno}{table_mrn_anno (total)}

\item{idx_seeds}{idx_seeds}

\item{round}{round}

\item{obj_mrn}{MRN object}

\item{table_ms2_network_filter}{table_ms2_network_filter}

\item{step_mrn}{step_mrn = 3 (MetDNA2/MRN2), = 1 (MetDNA3/MRN3)}

\item{topN_neighbor}{topN_neighbor = Inf (100 in MetDNA2)}

\item{lib_type_mrn}{"MRN3" or "MRN2"}

\item{thread}{thread = 4}
}
\value{
table_mrn_anno (after metabolite annotation)
}
\description{
getMetAnno
}
\examples{
table_mrn_anno <- getMetAnno(...)

}
