% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRN3getMrnNeighbor.R
\name{getMrnNeighbor}
\alias{getMrnNeighbor}
\title{getMrnNeighbor}
\usage{
getMrnNeighbor(id, step, graph)
}
\arguments{
\item{id}{A character string specifying the metabolite identifier.}

\item{step}{An integer indicating the number of steps to expand the neighborhood in the network.}

\item{graph}{The metabolic reaction network represented as an igraph object.}
}
\value{
A data frame containing the metabolite IDs and the corresponding steps in the network.
}
\description{
getMrnNeighbor
}
\examples{
# Assuming 'graph' is an igraph object representing the metabolic reaction network
# Get neighbors of metabolite with ID 'C03793' within 3 steps
getMrnNeighbor(id = 'C03793', step = 3, graph = MetDNA3::reaction_pair_network$version2$step2)

}
\seealso{
\code{\link{igraph::neighbors}}
}
