% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRN3preparation.R
\name{getMs1Data4Mrn3}
\alias{getMs1Data4Mrn3}
\title{getMs1Data4Mrn3}
\usage{
getMs1Data4Mrn3(wd = ".", ms1_filename = "data.csv", is_output = TRUE)
}
\arguments{
\item{wd}{Working directory (default is the current directory).}

\item{ms1_filename}{Filename of the MS1 data (default is 'data.csv').}

\item{is_output}{output ms1_data?}
}
\value{
A tibble containing the MS1 data in MRN3 format with columns 'name', 'mz', 'rt', and 'ccs'.
}
\description{
getMs1Data4Mrn3
}
\examples{
getMs1Data4Mrn3(wd = 'TO/YOUR/PATH/', ms1_filename = 'data.csv')

}
