% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRN3preparation.R
\name{getMs2Data4Mrn3}
\alias{getMs2Data4Mrn3}
\title{getMs2Data4Mrn3}
\usage{
getMs2Data4Mrn3(
  wd = ".",
  ms2_filename = "spectra.msp",
  is_output = TRUE,
  is_use_01_ms2 = TRUE
)
}
\arguments{
\item{wd}{Working directory (default is the current directory).}

\item{ms2_filename}{Filename of the MS2 spectra data (default is 'spectra.msp').}

\item{is_output}{output ms2_data?}

\item{is_use_01_ms2}{use MS2 in 01_result_initial_seed_annotation/00_intermediate_data/ms2}
}
\value{
A list containing MS2 spectra data in MRN3 format.
}
\description{
Converts MS2 spectra data to MRN3 format.
}
\examples{
getMs2Data4Mrn3(wd = 'TO/YOUR/PATH/', ms2_filename = 'spectra.msp')

}
