% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRN3preparation.R
\name{getTableSeedLong4Mrn3}
\alias{getTableSeedLong4Mrn3}
\title{getTableSeedLong4Mrn3}
\usage{
getTableSeedLong4Mrn3(
  wd = ".",
  seedAnno_filename = "01_result_initial_seed_annotation/ms2_match_annotation_result.csv",
  id_type = c("id_mrn", "id_kegg"),
  is_output = TRUE
)
}
\arguments{
\item{wd}{Working directory (default is the current directory).}

\item{seedAnno_filename}{Filename of the seed annotation file (default is '01_result_initial_seed_annotation/ms2_match_annotation_result.csv').}

\item{id_type}{ID conversion to "id_mrn" or "id_kegg"}

\item{is_output}{output seed table (long-format)?}
}
\value{
A tibble containing the long-format seed table for MRN3 annotation with columns 'feature_name' and 'id_mrn'.
}
\description{
Reads the seed annotation file and converts it to a long-format table for MRN3 annotation.
}
\examples{
getTableSeedLong4Mrn3(wd = 'TO/YOUR/PATH/', seedAnno_filename = '01_result_initial_seed_annotation/ms2_match_annotation_result.csv')

}
