% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRN3main.R
\name{MrnAnnoAlgo}
\alias{MrnAnnoAlgo}
\title{MrnAnnoAlgo}
\usage{
MrnAnnoAlgo(
  wd = ".",
  ms1_data,
  ms2_data,
  table_seed_long,
  instrument_type = c("Orbitrap", "TOF"),
  column = c("hilic", "rp"),
  method_lc = c("Amide12min", "Amide23min", "Other", "MetlinRP", "RP12min", "RP12min_v2"),
  polarity = c("positive", "negative"),
  lib_type_mrn = c("MRN3", "MRN2"),
  ex_step = c("0", "1", "2"),
  is_rt_pred_by_metdna2 = TRUE,
  adduct_for_MRN_anno = NULL,
  ms2_sim_method = c("dp", "hybrid"),
  thread = 4
)
}
\arguments{
\item{wd}{Working directory (default is the current directory).}

\item{ms1_data}{MS1 data in MRN3 format with columns 'name', 'mz', 'rt', and 'ccs'}

\item{ms2_data}{A list containing MS2 spectra data in MRN3 format (feature name matched with MS1).}

\item{table_seed_long}{A tibble containing the long-format seed table for MRN3 annotation with columns 'feature_name' and 'id_mrn'.}

\item{instrument_type}{instrument_type: "Orbitrap" or "TOF"}

\item{column}{LC type: "hilic" or "rp"}

\item{method_lc}{LC mwethod: 'Amide12min', 'Amide23min', 'Other', 'MetlinRP', 'RP12min', 'RP12min_v2'}

\item{polarity}{polarity: "positive" or "negative"}

\item{lib_type_mrn}{'MRN3', 'MRN2'}

\item{ex_step}{extension step: '0', '1', '2'}

\item{is_rt_pred_by_metdna2}{TRUE!}

\item{adduct_for_MRN_anno}{NULL or limited to some types}

\item{ms2_sim_method}{dp or hybrid}

\item{thread}{thread (default is 4).}
}
\value{
null
}
\description{
MrnAnnoAlgo
}
\examples{
MrnAnnoAlgo(
wd = 'TO/YOUR/PATH/',
ms1_data = ms1_data,
ms2_data = ms2_data,
table_seed_long = table_seed_long,
thread = 8)

}
