% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRN3preparation.R
\name{predictRtInMrn3}
\alias{predictRtInMrn3}
\title{predictRtInMrn3}
\usage{
predictRtInMrn3(
  wd,
  ms1_data,
  table_seed_long,
  md_mrn,
  column = c("hilic", "rp")
)
}
\arguments{
\item{wd}{Working directory}

\item{ms1_data}{ms1_data}

\item{table_seed_long}{table_seed_long}

\item{md_mrn}{md_mrn}

\item{column}{column}
}
\value{
"mrn_rt" (Named predicted RT vector of rownames(md_mrn))
}
\description{
predictRtInMrn3
}
\examples{
predictRtInMrn3(wd, ms1_data, table_seed_long, md_mrn, column = 'hilic')

}
