% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRN3rmRedun.R
\name{reconstructRtPredModel}
\alias{reconstructRtPredModel}
\title{reconstructRtPredModel}
\usage{
reconstructRtPredModel(result_df_rm_by_f, md_mrn, column, wd_output)
}
\arguments{
\item{result_df_rm_by_f}{'result_df_rm_by_f' of rmRedunInMetdna3 function}

\item{md_mrn}{'md_mrn' of rmRedunInMetdna3 function}

\item{column}{'column' of rmRedunInMetdna3 function}

\item{wd_output}{'wd_output' of rmRedunInMetdna3 function}
}
\value{
MRN RTs
}
\description{
reconstructRtPredModel
}
\examples{
reconstructRtPredModel(result_df_rm_by_f)
}
