% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRN3rmRedun_sameWithMetdna.R
\name{rmRedunInMrn3}
\alias{rmRedunInMrn3}
\title{rmRedunInMrn3}
\usage{
rmRedunInMrn3(result_df, is_only_confidence_assignment = FALSE)
}
\arguments{
\item{result_df}{annotation result data frame (all annotation results!)}

\item{is_only_confidence_assignment}{FALSE}
}
\value{
annotation result data frame after redundancy removal (with column: "group" & "confidence" !)
}
\description{
rmRedunInMrn3
}
\examples{
rmRedunInMrn3(result_df)

}
