% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRN3rmRedun.R
\name{rmRedunInMetdna3}
\alias{rmRedunInMetdna3}
\title{rmRedunInMetdna3}
\usage{
rmRedunInMetdna3(
  result_df,
  num_candi = 3,
  is_known_prioritization = T,
  wd_output
)
}
\arguments{
\item{result_df}{result_mrn_anno table (feature-metabolite pairs)}

\item{num_candi}{number of candidates (topN) for redundancy removal}

\item{is_known_prioritization}{priority: level3.1 (knowns) > level3.2 (unknowns) for redundancy removal}

\item{wd_output}{file.path(wd, '02_result_MRN_annotation')}
}
\value{
result_mrn_anno table with redundancy removal
}
\description{
rmRedunInMetdna3
}
